Added Service classes and servlets and filled in the Add method from Servlet to Dao
I filled in the info from Servlet to Dao for an Add method for the Employee Class.  I think that's how it will work I'm just not famliar with how the servlet uses it but hopefully what I did makes sense.  It should be the exact same thing for all the others everywhere I think.