Completed title workflow and added extra fields in the add method.
having issues displaying the addEmployee.jsp page now because for some reason it's not reading the department elements of the departments array list to access the deptNo of that object.  This is to populate a drop down list for the departments.  hopefully in the morning things will make more sense!